+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed `Params::is_constructor` to return the suggested value instead of `false`.
+
 ### Security
 
 ## [0.13.0] - 2026-09-07